+++ tests/testthat/test-oa_snowball.R
@@ -68,7 +68,7 @@ test_that("oa_snowball works for author orcids", {
   )
 
   nodes <- snowball_orcid$nodes
-  orcids_in <- lapply(nodes$author[nodes$oa_input], function(x) x$au_orcid)
+  orcids_in <- lapply(nodes$author[nodes$oa_input], function(x) x$orcid)
   either_orcid <- paste(orcids, collapse = "|")
 
   expect_true(is.list(snowball_orcid))